requirements.txt
awscli==0.7.0
bcrypt==3.1.7
boto3==1.10.37
botocore==1.13.50
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cryptography==3.3.2
docutils==0.15.2
idna==2.8
Jinja2==2.11.3
jmespath==0.9.4
jsonpickle==1.2
MarkupSafe==1.1.1
Naked==0.1.31
paramiko==2.6.0
Pillow==8.2.0
pyasn1==0.4.8
pycparser==2.19
pycryptodome==3.9.4
PyNaCl==1.3.0
python-dateutil==2.8.0
python-digitalocean==1.14.0
PyYAML==5.4
qrcode==6.1
requests==2.25.1
rsa==4.7
s3transfer==0.2.1
scp==0.13.2
shellescape==3.4.1
six==1.13.0
urllib3==1.25.11
